+++ gmd/Common/Private/FileStore.cs
@@ -32,7 +32,7 @@ class FileStore : IFileStore
         try
         {
             string json = JsonSerializer.Serialize(state, options);
-            if (!Try(out var e, Files.WriteAllText(path, json))) Asserter.FailFast(e.ErrorMessage);
+            if (!Try(out var e, () => File.WriteAllText(path, json))) Asserter.FailFast(e.ErrorMessage);
             cache[path] = state!;
         }
         catch (Exception e)
@@ -50,12 +50,12 @@ class FileStore : IFileStore
                 return (T)cached;
             }
 
-            if (!Files.Exists(path))
+            if (!File.Exists(path))
             {
                 Write(path, (T)Activator.CreateInstance(typeof(T))!);
             }
 
-            if (!Try(out var json, out var e, Files.ReadAllText(path))) throw Asserter.FailFast(e.ErrorMessage);
+            if (!Try(out var json, out var e, () => File.ReadAllText(path))) throw Asserter.FailFast(e.ErrorMessage);
             var state = JsonSerializer.Deserialize<T>(json);
             if (state == null)
             {